mendel_info bug
