+++ SynTool/chem/reaction_rules/manual/decompositions.py
