+++ SynTool/chem/reaction_rules/manual/transformations.py
