+++ SynTool/ml/networks/value.py
@@ -9,8 +9,8 @@ from torch.nn import Linear
 from torch.nn.functional import binary_cross_entropy_with_logits
 from torchmetrics.functional.classification import binary_recall, binary_specificity, binary_f1_score
 
-from Syntool.ml.networks.modules import MCTSNetwork
-from Syntool.utils.config import ConfigABC
+from SynTool.ml.networks.modules import MCTSNetwork
+from SynTool.utils.config import ConfigABC
 
 
 class SynthesabilityValueNetwork(MCTSNetwork, LightningModule, ABC):
