+++ SynTool/ml/training/__init__.py
@@ -1,4 +1,3 @@
-from .loading import *
 from .supervised import *
 from .preprocessing import ValueNetworkDataset, mol_to_pyg, MENDEL_INFO
 from .supervised import create_policy_dataset, run_policy_training
@@ -7,8 +6,6 @@ __all__ = [
     "ValueNetworkDataset",
     "mol_to_pyg",
     "MENDEL_INFO",
-    "load_policy_net",
-    "load_value_net",
     'create_policy_dataset',
     'run_policy_training'
 ]
