+++ SynTool/utils/config.py
@@ -10,7 +10,7 @@ from CGRtools.containers import MoleculeContainer, QueryContainer, ReactionConta
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any, Dict
-from Syntool.utils import path_type
+from SynTool.utils import path_type
 
 
 @dataclass
