--- synto_env.yml
@@ -1,230 +1,0 @@
-name: syntool_env
-channels:
-  - pytorch
-  - pyg
-  - nvidia
-  - conda-forge
-dependencies:
-  - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=2_kmp_llvm
-  - annotated-types=0.6.0=pyhd8ed1ab_0
-  - anyio=3.7.1=pyhd8ed1ab_0
-  - appdirs=1.4.4=pyh9f0ad1d_0
-  - arrow=1.3.0=pyhd8ed1ab_0
-  - backoff=2.2.1=pyhd8ed1ab_0
-  - beautifulsoup4=4.12.2=pyha770c72_0
-  - blas=2.116=mkl
-  - blas-devel=3.9.0=16_linux64_mkl
-  - blessed=1.19.1=pyhe4f9e05_2
-  - brotli-python=1.1.0=py311hb755f60_1
-  - bzip2=1.0.8=hd590300_5
-  - ca-certificates=2023.11.17=hbcca054_0
-  - cachecontrol=0.13.1=pyhd8ed1ab_0
-  - cachecontrol-with-filecache=0.13.1=pyhd8ed1ab_0
-  - cachy=0.3.0=pyhd8ed1ab_1
-  - certifi=2023.11.17=pyhd8ed1ab_0
-  - cffi=1.16.0=py311hb3a22ac_0
-  - charset-normalizer=3.3.2=pyhd8ed1ab_0
-  - cleo=2.1.0=pyhd8ed1ab_0
-  - click=8.1.7=unix_pyh707e725_0
-  - click-default-group=1.2.4=pyhd8ed1ab_0
-  - clikit=0.6.2=pyhd8ed1ab_2
-  - colorama=0.4.6=pyhd8ed1ab_0
-  - conda-lock=2.5.1=pyhd8ed1ab_0
-  - crashtest=0.4.1=pyhd8ed1ab_0
-  - croniter=1.4.1=pyhd8ed1ab_0
-  - cryptography=41.0.7=py311hcb13ee4_1
-  - cuda-cudart=11.8.89=0
-  - cuda-cupti=11.8.87=0
-  - cuda-libraries=11.8.0=0
-  - cuda-nvrtc=11.8.89=0
-  - cuda-nvtx=11.8.86=0
-  - cuda-runtime=11.8.0=0
-  - dateutils=0.6.12=py_0
-  - dbus=1.13.6=h5008d03_3
-  - deepdiff=6.7.1=pyhd8ed1ab_0
-  - distlib=0.3.7=pyhd8ed1ab_0
-  - dulwich=0.21.7=py311h459d7ec_0
-  - ensureconda=1.4.3=pyhd8ed1ab_0
-  - exceptiongroup=1.2.0=pyhd8ed1ab_0
-  - expat=2.5.0=hcb278e6_1
-  - fastapi=0.104.1=pyhd8ed1ab_0
-  - filelock=3.13.1=pyhd8ed1ab_0
-  - fsspec=2023.12.1=pyhca7485f_0
-  - gdown=4.7.1=pyhd8ed1ab_0
-  - gettext=0.21.1=h27087fc_0
-  - gitdb=4.0.11=pyhd8ed1ab_0
-  - gitpython=3.1.40=pyhd8ed1ab_0
-  - gmp=6.3.0=h59595ed_0
-  - gmpy2=2.1.2=py311h6a5fa03_1
-  - h11=0.14.0=pyhd8ed1ab_0
-  - html5lib=1.1=pyh9f0ad1d_0
-  - icu=73.2=h59595ed_0
-  - idna=3.6=pyhd8ed1ab_0
-  - importlib-metadata=7.0.0=pyha770c72_0
-  - importlib_metadata=7.0.0=hd8ed1ab_0
-  - inquirer=3.1.4=pyhd8ed1ab_0
-  - itsdangerous=2.1.2=pyhd8ed1ab_0
-  - jaraco.classes=3.3.0=pyhd8ed1ab_0
-  - jeepney=0.8.0=pyhd8ed1ab_0
-  - jinja2=3.1.2=pyhd8ed1ab_1
-  - joblib=1.3.2=pyhd8ed1ab_0
-  - keyring=24.3.0=py311h38be061_0
-  - ld_impl_linux-64=2.40=h41732ed_0
-  - libblas=3.9.0=16_linux64_mkl
-  - libcblas=3.9.0=16_linux64_mkl
-  - libcublas=11.11.3.6=0
-  - libcufft=10.9.0.58=0
-  - libcufile=1.8.1.2=0
-  - libcurand=10.3.4.101=0
-  - libcusolver=11.4.1.48=0
-  - libcusparse=11.7.5.86=0
-  - libexpat=2.5.0=hcb278e6_1
-  - libffi=3.4.2=h7f98852_5
-  - libgcc-ng=13.2.0=h807b86a_3
-  - libgfortran-ng=13.2.0=h69a702a_3
-  - libgfortran5=13.2.0=ha4646dd_3
-  - libglib=2.78.2=h783c2da_0
-  - libhwloc=2.9.3=default_h554bfaf_1009
-  - libiconv=1.17=h166bdaf_0
-  - liblapack=3.9.0=16_linux64_mkl
-  - liblapacke=3.9.0=16_linux64_mkl
-  - libnpp=11.8.0.86=0
-  - libnsl=2.0.1=hd590300_0
-  - libnvjpeg=11.9.0.86=0
-  - libsqlite=3.44.2=h2797004_0
-  - libstdcxx-ng=13.2.0=h7e041cc_3
-  - libuuid=2.38.1=h0b41bf4_0
-  - libxml2=2.11.6=h232c23b_0
-  - libzlib=1.2.13=hd590300_5
-  - lightning=2.1.2=pyhd8ed1ab_0
-  - lightning-cloud=0.5.39=pyhd8ed1ab_0
-  - lightning-utilities=0.10.0=pyhd8ed1ab_0
-  - llvm-openmp=15.0.7=h0cdce71_0
-  - markdown-it-py=3.0.0=pyhd8ed1ab_0
-  - markupsafe=2.1.3=py311h459d7ec_1
-  - mdurl=0.1.0=pyhd8ed1ab_0
-  - mkl=2022.1.0=h84fe81f_915
-  - mkl-devel=2022.1.0=ha770c72_916
-  - mkl-include=2022.1.0=h84fe81f_915
-  - more-itertools=10.1.0=pyhd8ed1ab_0
-  - mpc=1.3.1=hfe3b2da_0
-  - mpfr=4.2.1=h9458935_0
-  - mpmath=1.3.0=pyhd8ed1ab_0
-  - msgpack-python=1.0.7=py311h9547e67_0
-  - ncurses=6.4=h59595ed_2
-  - networkx=3.2.1=pyhd8ed1ab_0
-  - numpy=1.26.2=py311h64a7726_0
-  - openssl=3.2.0=hd590300_1
-  - ordered-set=4.1.0=pyhd8ed1ab_0
-  - orjson=3.9.10=py311h34b1e23_0
-  - packaging=23.2=pyhd8ed1ab_0
-  - pandas=2.1.3=py311h320fe9a_0
-  - pastel=0.2.1=pyhd8ed1ab_0
-  - pcre2=10.42=hcad00b1_0
-  - pexpect=4.8.0=pyh1a96a4e_2
-  - pip=23.3.1=pyhd8ed1ab_0
-  - pkginfo=1.9.6=pyhd8ed1ab_0
-  - platformdirs=3.11.0=pyhd8ed1ab_0
-  - poetry=1.7.1=linux_pyha804496_0
-  - poetry-core=1.8.1=pyhd8ed1ab_0
-  - poetry-plugin-export=1.6.0=pyhd8ed1ab_0
-  - psutil=5.9.5=py311h459d7ec_1
-  - ptyprocess=0.7.0=pyhd3deb0d_0
-  - pycparser=2.21=pyhd8ed1ab_0
-  - pydantic=2.1.1=pyhd8ed1ab_0
-  - pydantic-core=2.4.0=py311h46250e7_0
-  - pyg=2.4.0=py311_torch_2.1.0_cu118
-  - pygments=2.17.2=pyhd8ed1ab_0
-  - pyjwt=2.8.0=pyhd8ed1ab_0
-  - pylev=1.4.0=pyhd8ed1ab_0
-  - pyparsing=3.1.1=pyhd8ed1ab_0
-  - pyproject_hooks=1.0.0=pyhd8ed1ab_0
-  - pysocks=1.7.1=pyha2e5f31_6
-  - python=3.11.6=hab00c5b_0_cpython
-  - python-build=1.0.3=pyhd8ed1ab_0
-  - python-dateutil=2.8.2=pyhd8ed1ab_0
-  - python-editor=1.0.4=py_0
-  - python-fastjsonschema=2.19.0=pyhd8ed1ab_0
-  - python-installer=0.7.0=pyhd8ed1ab_0
-  - python-multipart=0.0.6=pyhd8ed1ab_0
-  - python-tzdata=2023.3=pyhd8ed1ab_0
-  - python_abi=3.11=4_cp311
-  - pytorch=2.1.1=py3.11_cuda11.8_cudnn8.7.0_0
-  - pytorch-cuda=11.8=h7e8668a_5
-  - pytorch-lightning=2.1.1=pyhd8ed1ab_0
-  - pytorch-mutex=1.0=cuda
-  - pytz=2023.3.post1=pyhd8ed1ab_0
-  - pyyaml=6.0.1=py311h459d7ec_1
-  - rapidfuzz=3.5.2=py311hb755f60_0
-  - readchar=4.0.5=pyhd8ed1ab_0
-  - readline=8.2=h8228510_1
-  - requests=2.31.0=pyhd8ed1ab_0
-  - requests-toolbelt=1.0.0=pyhd8ed1ab_0
-  - rich=13.7.0=pyhd8ed1ab_0
-  - ruamel.yaml=0.18.5=py311h459d7ec_0
-  - ruamel.yaml.clib=0.2.7=py311h459d7ec_2
-  - scikit-learn=1.3.2=py311hc009520_1
-  - scipy=1.11.4=py311h64a7726_0
-  - secretstorage=3.3.3=py311h38be061_2
-  - setuptools=68.2.2=pyhd8ed1ab_0
-  - shellingham=1.5.4=pyhd8ed1ab_0
-  - six=1.16.0=pyh6c4a22f_0
-  - smmap=5.0.0=pyhd8ed1ab_0
-  - sniffio=1.3.0=pyhd8ed1ab_0
-  - soupsieve=2.5=pyhd8ed1ab_1
-  - starlette=0.27.0=pyhd8ed1ab_0
-  - starsessions=1.3.0=pyhd8ed1ab_0
-  - sympy=1.12=pypyh9d50eac_103
-  - tbb=2021.10.0=h00ab1b0_2
-  - threadpoolctl=3.2.0=pyha21a80b_0
-  - tk=8.6.13=noxft_h4845f30_101
-  - tomli=2.0.1=pyhd8ed1ab_0
-  - tomlkit=0.12.3=pyha770c72_0
-  - toolz=0.12.0=pyhd8ed1ab_0
-  - torchmetrics=1.2.1=pyhd8ed1ab_0
-  - torchtriton=2.1.0=py311
-  - tqdm=4.66.1=pyhd8ed1ab_0
-  - traitlets=5.14.0=pyhd8ed1ab_0
-  - trove-classifiers=2023.11.29=pyhd8ed1ab_0
-  - types-python-dateutil=2.8.19.14=pyhd8ed1ab_0
-  - typing-extensions=4.8.0=hd8ed1ab_0
-  - typing_extensions=4.8.0=pyha770c72_0
-  - tzdata=2023c=h71feb2d_0
-  - urllib3=1.26.18=pyhd8ed1ab_0
-  - uvicorn=0.24.0.post1=py311h38be061_0
-  - virtualenv=20.25.0=pyhd8ed1ab_0
-  - wcwidth=0.2.12=pyhd8ed1ab_0
-  - webencodings=0.5.1=pyhd8ed1ab_2
-  - websocket-client=1.7.0=pyhd8ed1ab_0
-  - websockets=12.0=py311h459d7ec_0
-  - werkzeug=3.0.1=pyhd8ed1ab_0
-  - wheel=0.42.0=pyhd8ed1ab_0
-  - xz=5.2.6=h166bdaf_0
-  - yaml=0.2.5=h7f98852_2
-  - zipp=3.17.0=pyhd8ed1ab_0
-  - pip:
-      - adabelief-pytorch==0.2.1
-      - aiosignal==1.3.1
-      - attrs==23.1.0
-      - cachedmethods==0.1.4
-      - cgrtools==4.1.35
-      - custom-inherit==2.4.1
-      - frozenlist==1.4.0
-      - future==0.18.3
-      - jsonschema==4.20.0
-      - jsonschema-specifications==2023.11.2
-      - multipledispatch==1.0.0
-      - pillow==10.1.0
-      - protobuf==4.25.1
-      - py-mini-racer==0.6.0
-      - pypng==0.20220715.0
-      - ray==2.8.1
-      - referencing==0.31.1
-      - reportlab==4.0.7
-      - rpds-py==0.13.2
-      - structurefingerprint==2.1
-      - tabulate==0.9.0
-      - toyplot==1.0.3
-      - toytree==2.0.1
-prefix: /home/tagir/mambaforge/envs/syntool_env #TODO change
